backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
supabase==0.7.1
aiofiles==0.7.0
Pillow==9.5.0
